refactor: restructure configs into public/ and private/
- rules/ directory removed entirely
- configs/private/ holds shared building blocks (not exported)
- configs/public/ holds all package.json exports
- Merged split file pairs: rules/base/es6+rules/es6, rules/base/style+rules/style
- Merged react+react-hooks+react-a11y into one public/react.js
- private/base.js now composes the full shared rule stack so
  public/browser.js and public/node.js only express what's unique to them
