--- configs/formatting.js
@@ -1,3 +1,0 @@
-import formatting from '../rules/formatting.js';
-
-export default formatting;
