+++ configs/private/strict.js
