+++ configs/public/formatting.js
