+++ configs/public/react-formatting.js
