+++ configs/public/scripts.js
