+++ configs/public/typescript.js
