--- configs/react-formatting.js
@@ -1,3 +1,0 @@
-import reactFormatting from '../rules/react-formatting.js';
-
-export default reactFormatting;
