--- configs/typescript.js
@@ -1,3 +1,0 @@
-import typescript from '../rules/typescript.js';
-
-export default typescript;
