+++ index.js
@@ -1,15 +1,13 @@
 import {defineConfig} from 'eslint/config';
 import globals from 'globals';
 
-import base from './configs/base.js';
-import browser from './configs/browser.js';
-import node from './configs/node.js';
-import react from './configs/react.js';
-import typescript from './configs/typescript.js';
-import expensify from './configs/expensify.js';
+import browser from './configs/public/browser.js';
+import node from './configs/public/node.js';
+import react from './configs/public/react.js';
+import typescript from './configs/public/typescript.js';
+import expensify from './configs/public/expensify.js';
 
 const config = defineConfig([
-    ...base,
     ...browser,
     ...node,
     ...react,
